services/github-scraper/requirements.txt
# GitHub Issue Scraper Service
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.1
aioredis==2.0.1
PyJWT==2.10.1
cryptography==44.0.0
pyyaml==6.0.2
apscheduler==3.10.4

# Dev / Test
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
respx==0.22.0